# Get-DbaDatabaseState, speedups